+++ sync-for-commerce/CodatSyncCommerce/CodatSyncCommerceSDK.cs
@@ -83,9 +83,10 @@ namespace CodatSyncCommerce
         };
 
         private const string _language = "csharp";
-        private const string _sdkVersion = "2.1.2";
-        private const string _sdkGenVersion = "2.131.1";
+        private const string _sdkVersion = "2.2.0";
+        private const string _sdkGenVersion = "2.139.1";
         private const string _openapiDocVersion = "1.1";
+        private const string _userAgent = "speakeasy-sdk/csharp 2.2.0 2.139.1 1.1 Codat.Sync.Commerce";
         private string _serverUrl = "";
         private ISpeakeasyHttpClient _defaultClient;
         private ISpeakeasyHttpClient _securityClient;