+++ sync-for-commerce/RELEASES.md
@@ -93,3 +93,13 @@ Based on:
 - [csharp v2.1.2] sync-for-commerce
 ### Releases
 - [NuGet v2.1.2] https://www.nuget.org/packages/Codat.Sync.Commerce/2.1.2 - sync-for-commerce
+
+## 2023-09-29 00:20:08
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce.yaml
+- Speakeasy CLI 1.91.3 (2.139.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v2.2.0] sync-for-commerce
+### Releases
+- [NuGet v2.2.0] https://www.nuget.org/packages/Codat.Sync.Commerce/2.2.0 - sync-for-commerce